+++ src/rust/ide/view/src/top_buttons.rs
@@ -1,0 +1,3 @@
+pub mod close;
+pub mod common;
+mod fullscreen;
